Python-based  authentication & login system
The username and password are stored as keys and values, respectively, in a Python dictionary. Next, I entered the user's username and password using their input. Printing login success or Invalid username or password will be displayed if the user name and password do not match the corresponding values in our data. Otherwise, please print the relevant values.